rename AttributeQuerying to Attributes::Querying
